requirements.txt
python-snap7==1.3
Flask==3.0.3
flask-cors==4.0.1
